Added tests for SingleGameInstance
