+++ packages/flame/lib/game.dart
@@ -7,6 +7,7 @@ export 'src/game/mixins/fps_counter.dart';
 export 'src/game/mixins/game.dart';
 export 'src/game/mixins/has_collidables.dart';
 export 'src/game/mixins/loadable.dart';
+export 'src/game/mixins/single_game_instance.dart';
 export 'src/game/projector.dart';
 export 'src/game/transform2d.dart';
 export 'src/text.dart';
